refactored code to keep dry
